requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.4
python-dotenv==1.0.0
reportlab==4.0.0  
